Handle Event/Job location display better
Added method Locations to Event table class to get a selection of Location
records for an Event. The selection only includes locations actually used
for a job, so if the Event default location is not used by any job, that
location will not be included in the selection.

Clean up code in signup.py.get_job_data to make use of the new method.
Use event.locations() in calendar.event.