Display Cancelled events in Calendar
So that visitors are more informed, all events except those with a
status of 'Pending' and and those with'exclude from calendar' == 1
will be displayed on the Calendar. Event with a status other than
'Scheduled' will have the status displayed above the event name.

Events other than 'Scheduled' will display the status in a header
bar on the calendar/event page.