Restrict signup for events
This may not really be needed, but incase someone wants to display
events with a status other than 'Scheduled' in signups list, now
appropreate messages are displayed and users may not signup for
the event. They may cancel their commitment however if they wish.

Since non 'scheduled' events and past events are not shown by default
in the list, this really has no effect, but I fell better about it
and that's what really matters.