Better volunteer signup display
Create a better url for the  "voluteer" link in calendar_event.html.
If user does not have a role that would allow them to sign up
for any job for the event now send them to the default signup
list page so they don't get the page that says nothing was found.